(RE-828) Fix shebangs in apple packages
Our packages were using /usr/bin/env ruby and installing to
system ruby loadpath, which breaks horribly in a multi-ruby (rbenv, etc)
environment. This goes into the workdir and looks for any files dropped in
bin, and "seds" the shebang to /usr/bin/ruby. I would love to be using a
ruby approach to this instead of shelling out to sed, but the problem is
we've already set ownership on these files, almost exclusively to root, and
thus we need to sudo out.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>