Use templated spec file when building an rpm instead of making a new one
Currently we prepare the workdir for building an rpm (SPECS, SOURCES) with a
new spec file that we evaluate from an erb in source and the tarball that gets
built. This is despite the fact that for almost all projects, we evaluate the
spec file template at tarball creation time, shipping a perfectly usable spec
file in ext/redhat. We should instead use the spec file if it exists in
ext/redhat, so that we can eventually decouple the creation of tarballs and
rpms and isolate where templating actually happens.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>