(maint) Use gem_name when pushing
Previously if a project name did not match the gem name, it would not get
pushed automatically during a pl:ship_gem. This commit updates the task to use
the gem_name instead of project name to avoid this problem. gem_name defaults
to project name if it is not set.