Add support for using reassignments to redirect initial variable values
In some cases, we're going to want to populate a variable with the value of
another one, e.g. project => gem_name. If gem_name isn't populated, we want to
use the value of project instead. This isn't going to work without some form of
assignment. This can be supplied by the Pkg::Config::issue_reassignments
method, which simply reassigns the value of an old variable to a new one, to
support all our old and crufty terrible ways.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>