Remove more references to @build
There are a few remaining references to the old '@build' object lying around in
the packaging repo. This commit removes them and replaces them with their
corresponding library equivalent. Likewise, there are a few references left to
old utilities that have been 'librarified' and these are fixed up as well.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>