Archive params file in jenkins uber_build and pass to repo job
This commit updates the dynamically generated packaging and repo creation jobs
we create on the jenkins server to include the archival of the parameters file
that is sent up originally, and then the use of this parameters file in the
repo job. This is primarily to support the current puppetdb paradigm, where the
switching of PE build vs non-PE build is done via environment variable. The
environment variable causes the "build_pe" setting to be true in the build
properties file, which is used in the initial packaging job, so we package
'pe-puppetdb' instead of 'puppetdb.' However, this information is lost in the
repo job, which thinks the project under package is regular puppetdb. By
passing the properties file along the chain, we ensure that any environment
variable overrides, including PE vs non-PE build, are honored not only in
packaging, but in repo creation.