Update signing for el4 and el7 packages
Prior, we were very particular about the packages that we would sign.
Specifically, we would only sign el5, el6 and fedora packages. This was
great and all until we realized that we would love to add in el4 and
el7. It was at that point, that we became sad, because there was no
el4/el7 signing love. They were abandoned, left adrift in the world of
unsigned packages. This commit brings them into the loving arms of a
family they never had, making them a part of the big family of signed
packages.