Add a Gemfile and travis file for testing the packaging repo
We're going to be testing the packaging tools more thoroughly than we currently
are, which means we should be testing in travis for its easiness and
awesomeness. However, to ensure we have rake and rspec ready in the testing
environment, we need a Gemfile for bundler.

Yes, I said it. Bundler. Alright, bundler. You won.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>