(RE-1044) Add packaging job URL to downstream job url parameters
The downstream job URL supplied by the user has 'PACKAGE_BUILD_STATUS'
automatically appended to it as a URL parameter, so that triggered jobs can
act on this information. It would also be useful to users to have the original
packaging job URL appended to the downstream job also, so that any failures are
readily accessible from the callback job.

This commit makes it so, by creating appending PACKAGE_BUILD_URL to the url as
a parameter as well, containing the url of the packaging job. We move the
downstream job to being created last, so that we can be certain the packaging
job has been created and assigned a url before we insert said url into the
downstream job. Turns out it jenkins jobs don't have to be created in order of
downstream,upstream so this is fine.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>