Curl downstream job url with verbose
If the downstream job fails to trigger, the curl call will just display the
response code and no additional information. This is not helpful when trying to
debug failures. This commit adds -v to the curl call so that it prints a full
trace of its attempt to trigger the job.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>