(maint) Refactor, improve pe_ship.rake
Previously the pe:update_yum_repo task was hardcoded to update el and sles
repos. In a world where eos is a valid platform, those hardcoded assumptions
break down. This commit updates the task to use the final_mocks as the first
and second elements of the platform tag to iterate over. Arch is left off the
tag, because the hardlinking we do for noarch packages means that noarch
packages only have an i386 mock, even when they are built for both i386 and
x86_64.
This refactor also removes the find and pushd from the shell snippet. The find
is not necessary because we know the names of the platform tags ahead of time.
The pushd is not needed because createrepo takes a $dir argument.