(maint) Set a Pkg::Config variable for packaging_root
PACKAGING_ROOT is set as a constant when invoked using packaging.rake, but no
Pkg::Config variable previously existed to record this information. This commit
adds a packaging_root variable that mirror project_root in several ways.
Pkg::Config::packaging_root defaults to PACKAGING_ROOT if it set, or if not
set, it will be set to the parent directory of lib, which tends to be
ext/packaging.