(RE-1172) Reset @packaging_root after loading overrides
Previously packaging_root would be set during the initial rake invocation,
which would set it to a fully qualified path on the machine invoking it, which
would not exist on the jenkins instance when running as a jenkins job. This
would break the filtering of packaging templates from being templatized, which
would in turn make the jenkins jobs fail. This commit addresses that by making
it a special case, as was done with @project_root. packaging_root is defined,
and then after the overrides from the PARAMS_FILE have been loaded (which would
clobber @packaging_root), it is reset.