(RE-1138) support versions less than zero as RC
Currently, a project must be either tagged with 'rc' or have an odd minor
version and specify odd_even versioning to get into the RC repos. We have a
request to support a project versioned with major less than 1 as RC repos as
well. This commit adds a new version strategy called 'zero_based' in which a
project that has a major version of 0 will always be considered not-final. This
is opt-in, which prevents breaking any existing usage.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>