(maint) More refactoring to keep us from shipping on anything but a tag
I've cleaned up numerous helper methods under lib/packaging/util and
revised the logical flow between notifications, prompts, and validations.

- Validations should only confirm a condition, or otherwise fail.

- Notifications should only relay information across stdout or stderr
(they should not change the execution flow of the program).

- Prompts should relay some basic amount of information and ask a user
to take action based on the information relayed.

These are the general workflow for many of the rake tasks we run (take a repo,
confirm user intent to do something with it, do the thing that we said we would)
so I've tried to construct basic methods that can be recycled as we tease apart
the numerous rake tasks.