(maint) Extend require_library_or_fail to case where gem name != library name
Note that some gems are named something other than their actual library
name. E.g. there is a ruby gem called 'jira' that you would load with
'require jira', but that is not the gem we want. Instead we want the
gem called 'ruby-jira' which is also loaded with 'require 'jira'.
Because ruby. Yay!

Prior to this patch, the 'require_library_or_fail' utility method
assumed the gem name and the ruby name were the same. This supports
the case where they are different (hopefully leading to a more useful
error message). This patch also flips tickets.rake over to use
'require_library_or_fail'.