(maint) Fix specs to run even after a package bootstrap
Prior to this change, if the packaging repo was bootstrapped
(i.e. so that it lives inside another repo), and that
repo specified a value for tar_host, this one spec would
fail because it assumed that value was nil.

This change just stashes the value of tar_host, and sets it
to nil to match the spec test assumption, and then restores
the original value when it's done.