(RE-1535) Restore previous template behavior
Previously, debian templates would be automatically evaluated and
included in tarballs. This functionality was lost at some point. This
commit works around the issue by sourcing the template from the root if
it doesn't exist in the workdir. If it is not available in either
location it exits with a helpful message.