Deprecate mkdir_pr, replace uses with FileUtils.mkdir_p
mkdir_pr is equivalent to a FileUtils.mkdir_p invoked with an array of
directories to create. This commit deprecates the mkdir_pr utility
method, and replaces all current invocations to use the FileUtils method
instead. The method itself is updated to pass its args directly to
mkdir_p instead of looping over them, as mkdir_p can handle a list of
directories.