Deprecate ln, update calls to use FileUtils.ln
The utility ln method provided by packaging does very little to augment
the FileUtils version. Our utility reverses the arguments and passes
some options. This commit updates any current calls to ln to use
FileUtils.ln and pass in the force and verbose flags, as is already
done. In tests, it looks as though the utility method is never invoked,
probably because of how rake requires and includes FileUtils. It also
deprecates the utility method, for future removal.