Add cow_list method to Pkg::Config
Currently our debian changelogs are littered with inaccurate lists of
debian platforms. Updating all of those lists independently of the
build_defaults file seems like a duplication of work. This commit
addresses that by making available a simple method to get a space
separated list of debian platforms for a project.