(maint) Fix template rejection
Previously the expand_templates called is_a? without invoking it against
an object. This meant that it was always false and so nothing was ever
rejected. The intended behavior was that templates that were part of
packaging itself would be excluded from the templates for the package.
This commit updates the is_a? call to be invoked on the temp object.