(maint) Add tests for the template method in tar.rb
This commit adds two tests for the template method in
lib/packaging/tar.rb. The first test verifies that the template method
correctly handles the distinction of hash vs string and uses the target
key in the hash instead of stripping off the extname of the template.
The second test simply validates that if the template cannot be found,
an error will be raised.