(RE-1746) Enhance JIRA release ticket docs
This provides extensive information based on the confluence page to the actual
tickets themselves, with the addition of what I have learned working through a
release.

This is a step in continual improvement. One thing I would like to see later
added to this task is a way for other subtasks in the same series block tickets
ahead. For now I have just noted those as dependencies in the descriptions.