Return raw git describe for tags
Previously, the version returned by the packaging repo would *always*
have the last component when split on '-' removed (when there are a
certain number of elements), which works very well for standard commits
and for our rc tags, but doesn't work for maven timestamps. For example,
currently 0.3.2-20140507.175526-5 is stripped to 0.3.2-20140507.175526.
This commit addresses that by explicitly using the raw git describe for
tags, and only doing manipulation for commits.