(RE-1784) Update debian packaging to prefer tarball artifacts
Previously, debian would template all of the artifacts from the project,
independent of the templating that happens during the tar creation, and
then only use the tar artifacts if the debian templating had not
happened, which is very rare (only puppetdb?). This commit updates the
deb packaging to copy the debian packaging files from the tarball on top
of any that had already been templated in the prep_deb phase.
Unfortunately, this doesn't avoid the double templating.