(maint) Allow rpm.rake to fall through to other spec sources
Previously the tar inspection happened using the ex method, which raises
an exception and fails if the command itself fails. This doesn't allow
the logic to fail through to templating the specfile anew from the
project as intended. This commit updates the logic to use %x and test $?
explicitly for success?. This will allow the logic to fall through
before failing as it was intended to.