(RE-1746) Clarify smoke test packages a bit
The current description is meant to be edited to handle smoke testing
packages and choosing platforms that will be tested against.

This also isn't very clear on the wonky way in which we create
MSIs and how they apply to the smoke testing subtask. This clears
that up so there is less confusion for folks new to the release
process.