(RE-2116) Build gem in a tmpdir
Previously, the gem was built out of the project_root, which worked well
as long as you didn't want to manipulate any of the files in the gem,
like to burn in a version string. This commit updates the create_gem
method to first make a tmp directory, then copy all of the gem files
into it, and build the gem.