(RE-2157) Add support for arbitrary deb mirrors to packaging
This commit leverages the work in a recent change to the pbuilderrc to
allow for overriding the secondary mirror to build against. It adds a
new build parameter to Pkg::Config, deb_build_mirrors, which should be
an array of deb sources to inject into the build using the BUILDMIRROR
environment variable. The dist can be templated in using __DIST__.