(maint) Move dist and arch setting out of conditional
Previously dist and arch were set inside of the else case of an if
block, which should mean that it vanishes once the block ends. Due to
either an oddity of ruby's scoping or some other bug, this has not yet
bitten us. This commit moves it out of the else block and simply ends
the conditional after the fail call, as that will exit the program
anyway.