(maint) Update load_envvars test to correctly set env
Previously this test set the environment variable in the wrong scope.
When running rspec with order set to random (as travis does),
occasionally these tests would be run after the new set_cow_env tests,
which reset PE_VER to nil. Because the ENV setting for the load_envvars
test happened on test LOAD rather than test EXECUTION (because they were
outside of the it block, that test would occasionally fail depending on
load order. This commit moves the ENV setting into the it block to
ensure this does not happen.