(maint) Set project root before each test in 00_utils version tests
Previously project_root would only be set once, before all the version
tests ran. This breaks on random order rspec runs because the it blocks
can be run interleaved with other tests. This commit updates the before
all to be a before each to ensure that project_root is correctly set
before each of the version tests.