(RE-1040) Use PATHEXT to resolve executables
Tools like git on windows are actually named 'git.exe', so we need to
take into account path extensions when resolving them.

This commit uses the PATHEXT environment variable, which is standard on
Windows, but if it is absent for some strange reason, it defaults to a
reasonable set of extensions.