(RE-2222) Push gems to internal server on release
This commit adjusts the ship_gem utility to include a try for pushing
the completed gem to the stickler host, currently rubygems.delivery.puppetlabs.net.
The host name can be overwritten using the environment variable INTERNAL_GEM_HOST.
If it fails it puts an error for user to check if they can access the server and have stickler installed.