(CPR-50) Change order of build-data
This commit goes in and changes the fetch.rake task to allow for project
data to overwrite team data. Before it would fetch the project data then
team data from the build-data repo, overwriting anything that is the
same in project. Since we currently do not have any intersections
between project and team datas this should not break anything.