(CPR-52) Parameterize sign_rpms task
To better reuse some of the existing tasks, adding parameters is
necessary. One such example is the sign_rpms task, which previously
assumed all rpms to be signed would be rooted under pkg. This assumption
fails when the rpms are in a directory called repos, which will be the
case when signing repos. This commit addresses that by adding a
parameter to the sign_rpms task which defaults to the current behavior
of pkg, but can be set to 'repos' or anything else as needed.