(CPR-52) Parameterize retrieve rake task
The pl:jenkins:retrieve rake task is limited in that it only stages
whatever it fetches in the pkg directory. This is not always the desired
behavior, such as for signing repos, so this commit adjusts that
behavior to be parameterized and default to the current behavior. This
will allow repos to be locally staged as repos instead of in a 'pkg'
directory.