(CPR-59) Move git utilities into library
This commit goes through and takes all of the git based utilities in
00_utils.rake and moves them to their own ruby file in packaging. It
also goes through and finds every use of the old version and replaces it
with the new one. Finally it deprecates the old one so as to show a
warning message in the event the old one is still used. However it won't
break, just warn.