(maint) Deprecate gpg_name in favor of gpg_key
gpg_name and gpg_key are both used in our signing interactions, and
gpg_key is always what we want to use, as gpg_name (an email address),
isn't guaranteed unique in the same way that they key (if long id is
being used) is. This commit deprecated gpg_name in favor of gpg_key, and
updates the loading to automatically populate gpg_key with the contents
of gpg_name if gpg_key is not set.