(CPR-52) Add deb repo signing method
The nightly repo work is working on already created repos, so all the
sign repos method for debian needs to do is to update the reprepro
configs for the repos and re-export them with signing enabled. This
commit adds a sign_repos method to Pkg::Deb::Repo to accomplish this.