(maint) Migrate rsync_* into Pkg::Util::Net
This commit migrates the rsync_to and rsync_from methods from 00_utils
into Pkg::Util::Net. 00_utils now has deprecations for rsync_to and
rsync_from and will automatically dispatch to the new methods. The new
methods have an explicit parameter list, which was already being used in
all instances in packaging. As it is being moved from rake to pure ruby,
ex is used instead of sh in the invocation.