(CPR-52) Parameterize Pkg::Deb::Repo repo_config methods
In signing nightly repos, it is important to keep the old repos and the
signed repos and their configs separate. This means that certain
assumptions that were baked into the repo config generation no longer
apply. The source and target of these methods need to be dynamic so that
they can be used on signed repos or repos in any path. This commit
applies those changes to Pkg::Deb::Repo to accomplish this.