(CPR-54) Refacter lib/packaging/deb/repo.rb
This commit makes some minor refactors to Pkg::Deb::Repo for both
correctness and cleanliness. First, all instances of find_tool with an
error are replaced with check_tool, which generates an error
automatically when the tool cannot be found. Second, sh is replaced by
Pkg:Util::Execution.ex to remove the dependency on rake in the library.
Similarly, FileUtils is required and then directly referenced in mkdir_p
calls to avoid the dependency on 00_utils. Finally, repo_urls is now
set outside an if block so that there is no need to initialize it as an
empty array. This also adds a base_url method that returns the builds
url that is shared between repos and repo_configs.