(maint) Use Kernel.system for remote_ssh_cmd instead of ex
Because ex saves all output to a variable, commands that depend on user
input, such as pinentry/gpg-agent will not work as the screen will never
be rendered because the command hasn't completed yet. This commit
updates that call to use Kernel.system instead, which only returns true
or false and doesn't capture the output of the call. The current
implementation caused ships/signs to hang indefinitely during the
keychain passphrase section.