(maint) Migrate ex to Pkg::Util::Execution
This commit migrates ex from 00_utils to a new module called
Pkg::Util::Execution. It also adds another method called success?, which
wraps the behavior of $?.success?, with the bonus of being more easily
stubbed/mocked in other tests. Some basic tests are included for ex. The
implementation of ex is moved from %x to `` to enable testability of the
method.