(maint) Address order dependent test failures in Pkg::Util::Net
Because the net spec tests around ssh stub out ssh calls, there is no
guarantee that the $? object will be instantiated, which can cause tests
to fail if we try to get success? from nil. This commit updates the
tests to use the Pkg::Util::Execution.success? method, which is easily
stubbed out in tests. The implementation is also updates to use the same
method.