(CPR-52) Add nightly_repos rake tasks
This commit is the first incantation of a task set to sign and ship
nightly repos. The main task nightly_repos, handles getting the repos
signed remotely, and then also ships and generates repo configs. As
many of the tasks as possible are ruby calls instead of rake task
invocations.