(CPR-52) Add create_repos method to Pkg::Rpm::Repo
To mirror the debian counterpart, this commit adds a method to
Pkg::Rpm::Repo to create repos from a directory. It is more atomic than
the other rpm repo creation in that we don't rsync, ship, or generate
repo configs in this method, just run createrepo on all the dirs.