(maint) Add signing_server config parameter
With the nightly repos, having a signing server that is distinct from
the distribution server is essential. This is currently implemented in a
hack in enterprise-dist's package promotion by temporarily setting the
distribution server to the signing server, but for the nightly repo we
need to have both config settings populated. This commit adds
signing_server as a new config setting with a SIGNING_SERVER environment
variable override.