(CPR-52) Allow generated rpm repo configs to set gpgcheck=1
For the nightly repos, the rpm repos will be signed, so the repo configs
should set gpgcheck to 1. This commit adds that ability via a parameter
to the method. When signed is true, the gpgkey field is also set in the
repo config and points to the builds_server root and a file named after
the gpg_key used. This commit also turns this setting on for the nightly
repos.