(CPR-52) Add deploy_nightly_repos task
This commit adds a deploy_nightly_repos task which will be used to ship
the signed nightly repos to the public nightly repository. It retreives
the repos and repo_configs and then swaps out the distribution server
for the nightly server, to ensure that the repo configs are functional.
This task also creates a latest symlink for the project.