(CPR-63) Template mock configs
This commit adds in capabilities that check if a template for mock
configs are in place then creates the mock config from them. It then
uses the newly made mock configs in the usual method. This is intended
to be used instead of needing a new config generated for every pe
version.
It will also be helpful for building for a pe version that hasn't had
mock configs generated yet.