(maint) Make ignore-existing optional for Pkg::Util::Net.rsync_to
There are occasions where it is desirable for rsync_to to clobber
existing files, such as for updating a latest directory on a nightly
repository. Previously, this behavior wasn't possible without shelling
out to a custom rsync, so this commit adds a parameter to the rsync_to
method to allow ignore-existing to be excluded from the rsync flags. It
defaults to true, to maintain existing behavior as the default.