(CPR-64) Refactor latest linking for nightly repos
This commit updates the nightly repos workflow to maintain a static
$project-latest directory with a linked repos directory. The
repo_configs are generated to have static filenames and contents, to
ensure that users can simply add the repos and have their package
manager keep them up to date with $project's master branch. The repos
symlink in the latest directory is updated for each project when a new
version ships, and this should be totally transparent to users of the
latest symlink.
This also tweaks the rsync out to not use --ignore-existing, as
otherwise the -latest directory is never updated or replaced.