(maint) Update rsync_to to take an array of arbitrary flags
Previously rsync_to included a toggle to turn ignore-existing on or off,
which was sufficient for that use case. For the nightly repos,
--delete-after will also be added to the rsync_to call, so this commit
updates the method to simple take an arbitrary array of flags which are
conditionally appended to the default rsync flags. The default behavior
is still to have --ignore-existing included.