(CPR-64) Cleanup latest repo_configs on each ship
Previously, repo_configs for the -latest directories could include
platforms that no longer exist in the nightlies latest repo. This commit
updates the rsync call to be two separate calls, one for the -latest
directory to cleanup any old files, and one for just the $project, to
push out the new build.