(maint) Update rsync_to call in ship.rake
This commit updates the rsync_to in ship.rake to use the flags parameter
as an array to pass the excludes and ignore-existing flags. It also
collapses the two remote calls into the same retry_on_fail_block. If the
first call fails, we aren't going to make it to the second call anyway.