(CPR-68) Dynamically create windows build job
This commit adds in the ability to create a dynamic jenkins job that
builds the windows MSI. It takes the git bundle used to build the linux
packages, and uses that version to build the msi.

We know to build the msi if :build_msi is set to true in
ext/build_defaults.yaml. If it is, then we're going to squeeze the msi
build job in after the packaging job, but before the repo job. If there
is a downstream job specified, that will be kicked off by the repo job
as always. If we are not building an msi, then the build pipeline
proceeds as normal (i.e. packaging => repo => downstream)

The main reasoning behind adding this work into the packaging repo is to
make it easier to build MSI's more frequently, and to have the versions
of all things in the msi in version control. Prior to this, the refs for
hiera/facter/puppet/etc. were tracked in the jenkins job itself. This
update allows us to maintain that information in ext/project_data.yaml.