(CPR-68) Add template for msi jenkins job
This adds in the template for the msi job. If :build_msi => true, then
we use this template. The job is kicked off as a downstream to the
packaging job, and has the repo job as its downstream. Otherwise, we do
nothing with this template, and the repo job is downstream of packaging.