(maint) Shorten Jenkins job name
Since this dynamically created jenkins job has a longer name, it's
actually exceeding the maximum path length allowed on Windows (260
characters). This is causing the job to fail when it's cloning into
puppet-win32-ruby. In order to fix this, we're using the short sha
rather than the long sha to name the job. Reducing the path by about 20
characters seems to fix the problem. Don't worry, though. I'm sure this
will come back to bite us again soon.