(CPR-69) allow msi build to be overridden with env var
We may not always want to build the MSI when we're building puppet. This
is especially true when we're building just packages that are going to
be tested. Because the MSI takes a while to build, we want to make sure
that the MSI build job can be overridden if desired. If the MSI env var
is not set, the msi will be created.