(RE-3310) Update PE mock substitution
Previously the PE mock substitution tried to match the provided PE
version against the mocks listed in the build_defaults. This works as
long as the build_defaults file is up to date, but fails if the
build_defaults' mock list is out of date, as is often the case. This
commit updates the automation to use a regex instead of literal match.
This assumes we never have a platform version that is in the x.y format,
which is currently true and seems unlikely to change.