(maint) Update release tickets
There are a few updates the release tickets could benefit from.

First, part of making sure the code is up to date is making sure all
outside software referenced it at the correct version. For Puppet, this
means making sure the shas/tags used to build the MSI are correct. For
Puppet-Server, this means making sure the Puppet versions are correct.

Second, Finch has created another commit-to-ticket comparison tool that
is very helpful, and should be mentioned. This is the
release-inquisition (because no one expects the release inquisition!).
This will definitely save a lot of time with these two tasks.

And finally, we need to make sure we're updating dujour for
Puppet-Server, as well as PuppetDB.