Update reprepro invocation to capture output
This commit leverages the previous commit's new facility in remote_ssh_cmd for
capturing the output of an ssh command. We do so to scan the output for
"Skipping Inclusion" - a string that indicates reprepro has refused to add a
deb to the repository because one of the same NVR already exists. When reprepro
does this, it exits 0. We were previously not handling this case, assuming
success based on the exit code. If reprepro has skipped inclusion, we now fail,
as we should.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>