(RE-3597) Move invoke_task into Pkg::Util::RakeUtils
This commit moves the invoke_task method into Pkg::Util::RakeUtils,
where it belongs. It also deprecates the old incantation and adds a
require on rake to the rake_utils module, as rake may not have already
been loaded. Other invocations of invoke_task are updated to use the new
method.