(RE-3710) Update the msi name
With the push for all in one agent packages, we're renaming our puppet
msi. This rename is so that the msi will be more in sync with the rest
of our puppet agent packages. This work is needed to build the windows
installer now that https://tickets.puppetlabs.com/browse/PUP-3844 has
been merged.

This also accounts for a new naming scheme that has been introducted
into the wix build job. Since we're now naming the installer
puppet-$version-$arch.msi, we need to account for that in our build job.
We no longer need to add the version to the name of the package. That,
and we're parsing out the version from the original build kick off, so
we don't need to look into the repo on the builder to find the version
information